refactor(auth): construct credential stores once and split per-backend files
Address review feedback on the credential-store orchestration:

- Wrap the keyring + plaintext pair in a CredentialStores struct built
  once from the FloxHub base URL and config dir, with the orchestration
  functions (probe_source, persist_login_token, resolve_into,
  clear_invalid, remove_all) as methods. This unifies the logout path
  and removes the config_dir aliasing clone in the CLI entrypoint.
- Hoist the "Auth0 mode, outside the prompt-hook flow" condition into a
  single named predicate at the call site instead of threading is_auth0
  and is_hook booleans into the resolver. resolve_floxhub_token stays
  outside the gate and gates on auth mode alone.
- Fold the credential-source reporting into
  CredentialSource::describe_storage and share the plaintext notice
  between 'auth status' and the login fallback warning.
- Split credential_store.rs into per-backend files: keyring.rs,
  plaintext.rs, and mock.rs, with the shared trait, error type, and
  CredentialStores orchestration in mod.rs. Per-store tests move with
  their store; shared test scaffolding lives in a test_helpers module.
